% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schema.R
\name{show_schema}
\alias{show_schema}
\title{Access Plataforma +Brasil's documentation}
\usage{
show_schema(verbose = TRUE)
}
\arguments{
\item{verbose}{Should the function display messages and progress bar? Defaults
to \code{TRUE}.}
}
\description{
\code{show_schema()} downloads, unzips, and show in a web browser Plataforma +Brasil's
full database documentation, including its schema and other usefull information.
Docs are downloaded to a folder in the working directory.
}
\examples{
\dontrun{show_schema()}
}
