% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siconvr.R
\docType{package}
\name{siconvr-package}
\alias{siconvr}
\alias{siconvr-package}
\title{\pkg{siconvr} is a package designed to extract and clean data on federal targeted
transfers in Brazil from SICONV (Sistema de Gestao de Convenios e Contratos de Repasse,
in Portuguese).}
\description{
\code{siconvr} has one main function: \code{get_siconv()}
}
\details{
Under the hoods, this functions takes care of building the URLs for the
desired files; downloads, stores, and unzip files' contents; and read
the retrieved datasets as fast as possible using \code{data.tables}' \code{fread()}.

Users can select a wide range of datasets to download: data on proposals
submited to the Brazilian Federal Government requesting local investments;
details on these proposals; data on accepted projects, their evaluations;
and the total amounts transfered.

To check the function's documentation page, run \code{?get_siconv}.
}
\author{
\strong{Maintainer}: Fernando Meireles \email{fmeireles@ufmg.br} (\href{https://orcid.org/0000-0002-7027-2058}{ORCID})

Authors:
\itemize{
  \item Marcus Vinícius de Sá \email{mvdstorres@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
