% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siconv.R
\name{get_siconv}
\alias{get_siconv}
\title{Retrieve cleaned data from SICONV}
\usage{
get_siconv(dataset = NULL, encoding = "UTF-8", cache = TRUE)
}
\arguments{
\item{dataset}{A \code{string} indicating the dataset to retrieve. Valid options:
\itemize{
\item "programs" -- data on available programs maintained by the Federal Government;
\item "proposals" -- data on submited proposals to the Federal Government;
}

When not specified, \code{dataset} defaults to "programs".}

\item{encoding}{Encoding used to read raw files. Defaults to "UTF-8", but users
might need to change this option to other valid options (e.g., "latin1") if
data are returned with uncommon characters.}

\item{cache}{Should the function maintain a local copy of the data? To save
both time and SICONV's server resources, this defaults to \code{TRUE}.}
}
\description{
\code{get_siconv()} is the package's workhorse. It downloads, cleans, and
returns one of the several datasets maintained by SICONV on federal
targeted transfers in Brazil. Under the hoods, the function handles GET requests,
downloads and stores intermediary files, and reads as fast as possible the data
to a \link[tibble:tibble-package]{tibble} format convenient for use in analysis.
}
\examples{
\dontrun{df <- get_siconv(dataset = "proposals")}
}
